requirements.txt
psycopg2~=2.9.9
python-dotenv~=1.0.1
pip~=24.0
attrs~=23.2.0
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.3